utils: Move to package root
