+++ bind/data.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/airbloc/solgen/bind/language"
 	"github.com/airbloc/solgen/bind/template"
-	"github.com/airbloc/solgen/bind/utils"
+	"github.com/airbloc/solgen/utils"
 	"github.com/ethereum/go-ethereum/accounts/abi"
 )
