+++ bind/language/golang.go
@@ -5,7 +5,7 @@ import (
 	"regexp"
 
 	"github.com/airbloc/solgen/bind/template"
-	"github.com/airbloc/solgen/bind/utils"
+	"github.com/airbloc/solgen/utils"
 	"github.com/ethereum/go-ethereum/accounts/abi"
 )
